requirements.txt
easyocr==1.7.1
filelock==3.15.4
fsspec==2024.6.1
imageio==2.34.2
intel-openmp==2021.4.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.3.0
pyclipper==1.3.0.post5
pygame==2.6.0
pytesseract==0.3.10
python-bidi==0.4.2
PyYAML==6.0.1
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
tifffile==2024.6.18
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
